requirements.txt
pandas == 2.2.3
dotenv==0.9.9
Flask==3.1.0
psycopg2==2.9.10
python-dotenv==1.1.0
PyJWT==2.10.1